Monaspace: Drop many weights and adapt paths
SemiWide and ExtraWide might be nice in some contexts, but that is not
really the audience of Nerd Fonts I guess.

Lets start with few and we can always add more if there is really a
demand.

Simplify the paths; the old style stems from the time when we patched
all variants per source font. Nowerdays the fonts are few (3 per source)
and will almost always be installed together anyhow.

Shorten the naming a bit for, for example
ERROR: ====-< Family (ID 1)      too long (37 > 31): MonaspaceXenon Nerd Font Propo Medium

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>